backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
anthropic>=0.40.0
qdrant-client[fastembed]>=1.12.0
fastembed>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.9.0
httpx>=0.28.0
